Refactor bundle structure to match hashids-bundle conventions
- Rename SqidsBundle to RoukmouteSqidsBundle
- Rename SqidsExtension to RoukmouteSqidsExtension
- Move ArgumentResolver to ValueResolver directory
- Move config/services.php to Resources/config/services.php
- Rename phpspec.yml to phpspec.yml.dist
- Update configuration key from 'sqids' to 'roukmoute_sqids'

Closes #7
